+++ mozjs/src/jsglue.cpp
@@ -69,6 +69,50 @@ bool JS_ForOfIteratorNext(JS::ForOfIterator* iterator,
   return iterator->next(val, done);
 }
 
+// Expose templated functions for tracing
+
+void CallValueTracer(JSTracer* trc, JS::Heap<JS::Value>* valuep,
+                     const char* name) {
+  JS::TraceEdge(trc, valuep, name);
+}
+
+void CallIdTracer(JSTracer* trc, JS::Heap<jsid>* idp, const char* name) {
+  JS::TraceEdge(trc, idp, name);
+}
+
+void CallObjectTracer(JSTracer* trc, JS::Heap<JSObject*>* objp,
+                      const char* name) {
+  JS::TraceEdge(trc, objp, name);
+}
+
+void CallStringTracer(JSTracer* trc, JS::Heap<JSString*>* strp,
+                      const char* name) {
+  JS::TraceEdge(trc, strp, name);
+}
+
+void CallScriptTracer(JSTracer* trc, JS::Heap<JSScript*>* scriptp,
+                      const char* name) {
+  JS::TraceEdge(trc, scriptp, name);
+}
+
+void CallFunctionTracer(JSTracer* trc, JS::Heap<JSFunction*>* funp,
+                        const char* name) {
+  JS::TraceEdge(trc, funp, name);
+}
+
+void CallUnbarrieredObjectTracer(JSTracer* trc, JSObject** objp,
+                                 const char* name) {
+  js::UnsafeTraceManuallyBarrieredEdge(trc, objp, name);
+}
+
+void CallObjectRootTracer(JSTracer* trc, JSObject** objp, const char* name) {
+  JS::TraceRoot(trc, objp, name);
+}
+
+void CallValueRootTracer(JSTracer* trc, JS::Value* valp, const char* name) {
+  JS::TraceRoot(trc, valp, name);
+}
+
 // These functions are only intended for use in testing,
 // to make sure that the Rust implementation of JS::Value
 // agrees with the C++ implementation.
@@ -115,7 +159,7 @@ JSLinearString* AtomToLinearString(JSAtom* atom) {
   return JS::AtomToLinearString(atom);
 }
 
-/* These types are using maybe, so we manually unwrap it in these wrappers */
+// These types are using maybe so we manually unwrap them in these wrappers
 
 bool FromPropertyDescriptor(JSContext* cx,
                             JS::Handle<JS::PropertyDescriptor> desc_,
