+++ mozjs/src/jsglue.hpp
@@ -47,9 +47,10 @@
 #include "jsapi.h"
 #include "jsfriendapi.h"
 
-// Reexport some functions that are marked inline.
 
 namespace glue {
+
+// Reexport some functions that are marked inline.
 
 bool JS_Init();
 
@@ -58,11 +59,17 @@ void DeleteRealmOptions(JS::RealmOptions* options);
 JS::OwningCompileOptions* JS_NewOwningCompileOptions(JSContext* cx);
 void DeleteOwningCompileOptions(JS::OwningCompileOptions* optiosn);
 
+JS::shadow::Zone* JS_AsShadowZone(JS::Zone* zone);
+
+JS::CallArgs JS_CallArgsFromVp(unsigned argc, JS::Value* vp);
+
 void JS_StackCapture_AllFrames(JS::StackCapture*);
 void JS_StackCapture_MaxFrames(uint32_t max, JS::StackCapture*);
 void JS_StackCapture_FirstSubsumedFrame(JSContext* cx,
                                         bool ignoreSelfHostedFrames,
                                         JS::StackCapture*);
+
+// Reexport some methods
 
 bool JS_ForOfIteratorInit(
     JS::ForOfIterator* iterator, JS::HandleValue iterable,
@@ -70,9 +77,27 @@ bool JS_ForOfIteratorInit(
 bool JS_ForOfIteratorNext(JS::ForOfIterator* iterator,
                           JS::MutableHandleValue val, bool* done);
 
-JS::shadow::Zone* JS_AsShadowZone(JS::Zone* zone);
+// Expose templated functions for tracing
 
-JS::CallArgs JS_CallArgsFromVp(unsigned argc, JS::Value* vp);
+void CallValueTracer(JSTracer* trc, JS::Heap<JS::Value>* valuep,
+                     const char* name);
+void CallIdTracer(JSTracer* trc, JS::Heap<jsid>* idp, const char* name);
+void CallObjectTracer(JSTracer* trc, JS::Heap<JSObject*>* objp,
+                      const char* name);
+void CallStringTracer(JSTracer* trc, JS::Heap<JSString*>* strp,
+                      const char* name);
+void CallScriptTracer(JSTracer* trc, JS::Heap<JSScript*>* scriptp,
+                      const char* name);
+void CallFunctionTracer(JSTracer* trc, JS::Heap<JSFunction*>* funp,
+                        const char* name);
+void CallUnbarrieredObjectTracer(JSTracer* trc, JSObject** objp,
+                                 const char* name);
+void CallObjectRootTracer(JSTracer* trc, JSObject** objp, const char* name);
+void CallValueRootTracer(JSTracer* trc, JS::Value* valp, const char* name);
+
+// These functions are only intended for use in testing,
+// to make sure that the Rust implementation of JS::Value
+// agrees with the C++ implementation.
 
 void JS_ValueSetBoolean(JS::Value* value, bool x);
 bool JS_ValueIsBoolean(const JS::Value* value);
@@ -98,36 +123,16 @@ size_t GetLinearStringLength(JSLinearString* s);
 uint16_t GetLinearStringCharAt(JSLinearString* s, size_t idx);
 JSLinearString* AtomToLinearString(JSAtom* atom);
 
-bool FromPropertyDescriptor(JSContext* cx,
-                            JS::Handle<JS::PropertyDescriptor> desc,
-                            JS::MutableHandle<JS::Value> vp);
-bool JS_GetOwnPropertyDescriptorById(
-    JSContext* cx, JS::HandleObject obj, JS::HandleId id,
-    JS::MutableHandle<JS::PropertyDescriptor> desc, bool* isNone);
-bool JS_GetOwnPropertyDescriptor(JSContext* cx, JS::HandleObject obj,
-                                 const char* name,
-                                 JS::MutableHandle<JS::PropertyDescriptor> desc,
-                                 bool* isNone);
-bool JS_GetOwnUCPropertyDescriptor(
-    JSContext* cx, JS::HandleObject obj, const char16_t* name, size_t namelen,
-    JS::MutableHandle<JS::PropertyDescriptor> desc, bool* isNone);
-bool JS_GetPropertyDescriptorById(
-    JSContext* cx, JS::HandleObject obj, JS::HandleId id,
-    JS::MutableHandle<JS::PropertyDescriptor> desc,
-    JS::MutableHandleObject holder, bool* isNone);
-bool JS_GetPropertyDescriptor(JSContext* cx, JS::HandleObject obj,
-                              const char* name,
-                              JS::MutableHandle<JS::PropertyDescriptor> desc,
-                              JS::MutableHandleObject holder, bool* isNone);
-bool JS_GetUCPropertyDescriptor(JSContext* cx, JS::HandleObject obj,
-                                const char16_t* name, size_t namelen,
-                                JS::MutableHandle<JS::PropertyDescriptor> desc,
-                                JS::MutableHandleObject holder, bool* isNone);
-bool SetPropertyIgnoringNamedGetter(JSContext* cx, JS::HandleObject obj,
-                                    JS::HandleId id, JS::HandleValue v,
-                                    JS::HandleValue receiver,
-                                    JS::Handle<JS::PropertyDescriptor> ownDesc,
-                                    JS::ObjectOpResult& result);
+// These types are using maybe so we manually unwrap them in these wrappers
+
+bool FromPropertyDescriptor(JSContext *cx, JS::Handle<JS::PropertyDescriptor> desc, JS::MutableHandle<JS::Value> vp);
+bool JS_GetOwnPropertyDescriptorById(JSContext* cx, JS::HandleObject obj, JS::HandleId id, JS::MutableHandle<JS::PropertyDescriptor> desc, bool* isNone);
+bool JS_GetOwnPropertyDescriptor(JSContext* cx, JS::HandleObject obj, const char* name, JS::MutableHandle<JS::PropertyDescriptor> desc, bool* isNone);
+bool JS_GetOwnUCPropertyDescriptor(JSContext* cx, JS::HandleObject obj, const char16_t* name, size_t namelen, JS::MutableHandle<JS::PropertyDescriptor> desc, bool* isNone);
+bool JS_GetPropertyDescriptorById(JSContext* cx, JS::HandleObject obj, JS::HandleId id, JS::MutableHandle<JS::PropertyDescriptor> desc, JS::MutableHandleObject holder, bool* isNone);
+bool JS_GetPropertyDescriptor(JSContext* cx, JS::HandleObject obj, const char* name, JS::MutableHandle<JS::PropertyDescriptor> desc, JS::MutableHandleObject holder, bool* isNone);
+bool JS_GetUCPropertyDescriptor(JSContext* cx, JS::HandleObject obj, const char16_t* name, size_t namelen, JS::MutableHandle<JS::PropertyDescriptor> desc, JS::MutableHandleObject holder, bool* isNone);
+bool SetPropertyIgnoringNamedGetter(JSContext* cx, JS::HandleObject obj, JS::HandleId id, JS::HandleValue v, JS::HandleValue receiver, JS::Handle<JS::PropertyDescriptor> ownDesc, JS::ObjectOpResult& result);
 
 bool CreateError(JSContext* cx, JSExnType type, JS::HandleObject stack,
                  JS::HandleString fileName, uint32_t lineNumber,
