+++ rust-mozjs/src/gc/mod.rs
@@ -1,0 +1,2 @@
+pub mod custom;
+pub mod root;
