+++ rust-mozjs/src/lib.rs
@@ -54,6 +54,7 @@ pub mod rust;
 mod consts;
 pub mod conversions;
 pub mod error;
+pub mod gc;
 pub mod glue;
 pub mod panic;
 pub mod typedarray;
